[TASK] Setup more users, some UI fresh-up, add output for Behat tests

+++ Classes/Acme/DemoServer/Controller/Test/UserController.php
@@ -58,7 +58,8 @@ class UserController extends \TYPO3\Flow\Mvc\Controller\ActionController {
 	public function resetAction() {
 		$users = $this->userRepository->findAll();
 		foreach ($users as $user) {
-			if ($user->getUsername() !== 'admin') {
+			// FIXME Find another way to keep static setup data (e.g. source property)
+			if ($user->getUsername() !== 'admin' && $user->getUsername() !== 'user1' && $user->getUsername() !== 'user2') {
 				$this->userRepository->remove($user);
 			}
 		}
